refactor: rename notice modal
