+++ main.ts
@@ -16,7 +16,7 @@ import PeersModal from "./modals/peersModal";
 import { CHANGE_PRESENCE_EVENT, ChangePresenceEventDto } from "./events/changePresenceEvent";
 import { CHANGE_SETTING_EVENT, ChangeSettingEventDto } from "./events/changeSettingEvent";
 import { addCopyFunctionToDocumentKeyProperty } from "./view/controllYorkieDocumentKeyProperty";
-import NoticeModal from "./modals/NoticeModal";
+import NoticeModal from "./modals/noticeModal";
 
 
 const USER_EVENTS_LIST = ['input', 'delete', 'move', 'undo', 'redo', 'set'];
